В проект добавлены модули + мелкие правки
