+++ js/form.js
@@ -1,0 +1,14 @@
+/* функция проверяет максимальную длину заданной строки, возвращает true или false */
+/*
+function hasAllowedStringLength(string, maxLength) {
+  if (typeof string === 'string') {
+    if (string.length <= maxLength) {
+      return true;
+    }
+    return false;
+  }
+  throw new Error ('Введены некорректные данные');
+}
+
+hasAllowedStringLength('Стас, ты молодец!', 17);
+*/
